requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
pydantic==2.6.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
black==23.12.1
ruff==0.1.6
requests==2.31.0
PyYAML==6.0.1 